LayoutTests/platform/glib/imported/w3c/web-platform-tests/css/css-break/hit-test-transformed-inline-expected.txt

xx
xx

FAIL First x assert_equals: expected "span" but got "firstchild"
FAIL First child assert_equals: expected "firstchild" but got "mc"
FAIL Second child assert_equals: expected "secondchild" but got "span"
FAIL Last x assert_equals: expected "span" but got "mc"
